Move JsonStream inline implementation to inline .h
